chore(e2e): #112 phase 1 — merge zeitraster + wochentage into time-grid.spec.ts under describe.serial
Both former specs PUT the same `/api/v1/schools/:id/time-grid`
endpoint which is implemented as `deleteMany + createMany` over the
full TimeGrid (school-time-grid.service.ts:82-104). Running them in
parallel inside the chromium project on different workers produces
this race:

  Worker A (zeitraster): PUT { periods:[P1, P2_marker_A], schoolDays:[MO] }
  Worker B (wochentage): PUT { periods:[P1], schoolDays:[MO, SA] }
  tx_B commits LAST → DB has [P1] (no marker) → Worker A's API
  readback fails: "period with label E2E-NNNNNN .toBeDefined()" → flake.

This was the exact failure mode that turned PR #109 CI red on
2026-05-18 (issue #87 PR, even though that PR didn't touch the
time-grid surface). PR #54/#55 had previously chromium-only-skipped
ZEIT-01 to dodge an inter-project variant of the same race, but the
intra-chromium race survived because wochentage.spec.ts had no skip
and writes to the same singleton.

Phase 1 fix from Issue #112: merge BOTH writers into one describe
block under `test.describe.configure({ mode: 'serial' })`. Playwright
guarantees serial-mode tests run on ONE worker sequentially → no race
possible. The chromium-only-skip persists for inter-project safety
(firefox + mobile would still race the singleton).

This is NOT the "throwaway-school" architecture sketched in the
original #112 plan — that requires relaxing `Person.keycloakUserId`
from `@unique` to `@@unique([keycloakUserId, schoolId])` so one KC
user can own a Person per throwaway school. Schema change is bigger,
tracked as future #112 phase.

Both FIXME comments reference #112 per CLAUDE.md D4 so the
serial-mode workaround is traceable.

Coverage tradeoff: WOCH-01 previously ran on chromium AND firefox
(no chromium-only-skip in wochentage.spec.ts). After merge it inherits
the chromium-only-skip. Acceptable risk — the WOCH-01 toggle behavior
is shadcn UI + REST round-trip, neither browser-engine-sensitive.

Local: 3/3 pass on desktop chromium (10.1s, 1 worker), 3/3 skip
cleanly on desktop-firefox.

Refs #112.
